Move old utils, add new util method yield_inst
